+++ front/src/views/service/service-apply/clb/index.module.scss
@@ -1,0 +1,154 @@
+.addClbConfigureSideslider {
+  .applyClbFormContainer {
+    background-color: #f5f7fa;
+    margin-top: 0;
+    padding: 24px !important;
+  }
+
+  .formCardContainer {
+    margin-bottom: 16px;
+
+    &:last-of-type {
+      margin-bottom: 52px;
+    }
+
+    :global(.common-card-content) {
+      width: 100% !important;
+    }
+  }
+
+  .flexRow {
+    display: flex;
+
+    :global(.bk-form-item) {
+      margin-right: 32px;
+
+      &:last-of-type {
+        margin-right: 0;
+        flex: 1;
+      }
+    }
+  }
+
+  .componentWithPreview {
+    display: flex;
+    align-items: center;
+
+    .previewBtn {
+      margin-left: 16px;
+    }
+  }
+
+  .simpleShowContainer {
+    display: flex;
+    height: 40px;
+    align-items: center;
+    margin-bottom: 16px;
+
+    .label {
+      margin-right: 5px;
+    }
+
+    .value {
+      margin-left: 10px;
+    }
+
+    .bkhcm-icon-prompt {
+      margin-left: 7px;
+      color: #c4c6cc;
+      cursor: pointer;
+    }
+  }
+
+  .sliderWrap {
+    margin-bottom: 16px;
+
+    :global(.bk-slider) {
+      width: 50vw;
+
+      .bk-slider-runway {
+        left: 4px;
+
+        .bk-slider-button {
+          z-index: 1;
+        }
+      }
+
+      .bk-slider-input {
+        margin-left: 45px;
+
+        .input-item {
+          width: 120px;
+        }
+      }
+    }
+
+    .sliderUnitSuffix {
+      width: 55px;
+      height: 32px;
+      padding: 0 8px;
+      border: 1px solid #c4c6cc;
+      border-left-style: none;
+      border-radius: 2px 0 0 2px;
+      background-color: #fafbfd;
+      box-sizing: border-box;
+      text-align: center;
+      user-select: none;
+    }
+  }
+
+  :global(.bk-sideslider-footer) {
+    justify-content: space-between;
+
+    .applyClbBottomBar {
+      position: static;
+      padding: 0;
+      width: auto;
+      height: auto;
+      display: inline-block;
+      border-top: 0px solid #dcdee5;
+
+      .infoWrap {
+        position: relative;
+        padding: 0 24px;
+        display: flex;
+        align-items: center;
+
+        .label {
+          margin-right: 5px;
+
+          &.hasTips {
+            border-bottom: 1px dashed #979ba5;
+            cursor: pointer;
+          }
+        }
+
+        .value {
+          margin-left: 5px;
+          color: #ff9c01;
+          font-weight: 700;
+
+          .unit {
+            margin-left: 5px;
+            font-size: 14px;
+          }
+
+          .number {
+            font-size: 18px;
+          }
+        }
+
+        &:not(:first-of-type)::before {
+          content: '';
+          position: absolute;
+          left: 0;
+          top: 50%;
+          width: 1px;
+          height: 16px;
+          transform: translateY(-50%);
+          background-color: #dcdee5;
+        }
+      }
+    }
+  }
+}
